fix: sort failed

+++ src/main/java/org/cardanofoundation/explorer/api/model/response/pool/projection/PoolListProjection.java
@@ -22,7 +22,7 @@ public interface PoolListProjection {
 
   Double getSaturation();
 
-  Integer getVotingPower();
+  Double getVotingPower();
 
   Double getGovernanceParticipationRate();
 }
